requirements.txt
alabaster==0.7.12
apipkg==1.5
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.12.5
cffi==1.14.4
cfgv==3.2.0
chardet==4.0.0
coverage==5.3
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.10
detect-secrets==0.14.3
distlib==0.3.1
docutils==0.16
entrypoints==0.3
execnet==1.7.1
filelock==3.0.12
flake8==3.8.4
identify==1.5.10
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.2
jedi==0.18.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
markdown-it-py==0.5.8
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
myst-parser==0.12.10
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
nodeenv==1.5.0
notebook==6.1.6
packaging==20.8
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
pre-commit==2.9.3
prometheus-client==0.9.0
prompt-toolkit==3.0.8
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyGithub==1.53
Pygments==2.7.3
PyJWT==2.0.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.1
pytest-forked==1.3.0
pytest-mock==3.4.0
pytest-xdist==2.2.0
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.25.1
Send2Trash==1.5.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.3.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
terminado==0.9.1
testpath==0.4.4
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.2
virtualenv==20.2.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
